LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-button-element/button-events-expected.txt


PASS The submit event must be fired when click a button in submit status 
PASS The reset event must be fired when click a button in reset status 
FAIL The show event must be fired when click a button in menu status assert_equals: The button type should be 'menu'. expected "menu" but got "submit"
